summarise_backend/requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
importlib-metadata==4.11.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.21.5
opencv-python==4.1.2.30
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
pytesseract==0.3.9
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
